requirements.txt
APScheduler==3.9.1
lxml
pip==20.0.2
requests          
simplejson        
tornado==6.1
websocket-client==1.3.3
websockets==10.2
python-dotenv==0.20.0
redis
